requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0
mpmath>=1.2.0

# Configuration
PyYAML>=6.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=7.6.0

# Development tools (optional)
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0